requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
cloudinary==1.41.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
